Assignment 2

+++ src/main/resources/schema.sql
@@ -2,6 +2,6 @@ create table if not exists item (
     id identity,
     name varchar(50) not null,
     yearCreated int not null,
-    price float(5) not null,
+    price int not null,
     brand varchar(50) not null
 );
